@@ -27,27 +27,24 @@
   "private": false,
   "main": "output/index.js",
   "peerDependencies": {
-    "react": ">= 16.8.x < 18"
+    "react": ">= 16.8.x"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/enzyme": "^3.10.8",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.2.0",
-    "auto-changelog": "^2.2.1",
-    "cross-env": "^7.0.2",
-    "enzyme": "^3.11.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.0",
+    "@types/node": "^16.11.68",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "auto-changelog": "^2.4.0",
+    "cross-env": "^7.0.3",
     "get-folder-size": "^2.0.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-scripts": "4.0.0",
-    "react-test-renderer": "^17.0.1",
-    "typescript": "^4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-test-renderer": "^18.2.0",
+    "typescript": "^4.8.4",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
